Move the `wcdl/radio-helper-text` directory into the `components` directory and rename related CSS.
