+++ js/src/components/free-listings/choose-audience-section/choose-audience-section.js
@@ -10,7 +10,7 @@ import { useAdaptiveFormContext } from '~/components/adaptive-form';
 import AppRadioContentControl from '~/components/app-radio-content-control';
 import Section from '~/wcdl/section';
 import Subsection from '~/wcdl/subsection';
-import RadioHelperText from '~/wcdl/radio-helper-text';
+import RadioHelperText from '~/components/radio-helper-text';
 import SupportedCountrySelect from '~/components/supported-country-select';
 import VerticalGapLayout from '~/components/vertical-gap-layout';
 import './choose-audience-section.scss';
