+++ js/src/components/free-listings/choose-audience-section/choose-audience-section.scss
@@ -1,5 +1,5 @@
 .gla-choose-audience-section {
-	.wcdl-radio-helper-text {
+	.gla-radio-helper-text {
 		font-style: normal;
 	}
