+++ js/src/components/radio-helper-text/index.scss
@@ -1,4 +1,4 @@
-.wcdl-radio-helper-text {
+.gla-radio-helper-text {
 	font-size: $helptext-font-size;
 	font-style: italic;
 	font-weight: 400;
