+++ js/src/components/shipping-rate-section/shipping-rate-section.js
@@ -10,7 +10,7 @@ import { createInterpolateElement } from '@wordpress/element';
 import { useAdaptiveFormContext } from '~/components/adaptive-form';
 import Section from '~/wcdl/section';
 import AppRadioContentControl from '~/components/app-radio-content-control';
-import RadioHelperText from '~/wcdl/radio-helper-text';
+import RadioHelperText from '~/components/radio-helper-text';
 import AppDocumentationLink from '~/components/app-documentation-link';
 import VerticalGapLayout from '~/components/vertical-gap-layout';
 import FlatShippingRatesInputCards from './flat-shipping-rates-input-cards';
